Refactor Header Code based on Example Implementation
- Restructure app structure to feature-based (per Johannes example)
- Include index.js files for API-like exports
- Refactor Header.js code to eliminate custom <Navigation> style
- Replace Header Tests with Johannes examples, which check for hash location
