+++ src/app/App.js
@@ -1,0 +1,33 @@
+import React from 'react';
+import { Switch, Route } from 'react-router-dom';
+import { ThemeProvider } from 'styled-components';
+import { GlobalStyle, theme } from '../style';
+
+import Header from '../common/header';
+import Search from '../pages/Search';
+import Home from '../pages/Home';
+import Footer from '../components/Footer';
+
+import Wrapper from './App.style';
+
+function App() {
+  return (
+    <Wrapper>
+      <ThemeProvider theme={theme}>
+        <GlobalStyle />
+        <Header />
+        <Switch>
+          <Route path="/search">
+            <Search />
+          </Route>
+          <Route path="/">
+            <Home />
+          </Route>
+        </Switch>
+        <Footer />
+      </ThemeProvider>
+    </Wrapper>
+  );
+}
+
+export default App;
