+++ src/app/App.style.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const Wrapper = styled.div`
+  margin: auto;
+  max-width: 1440px;
+`;
+
+export default Wrapper;
