+++ src/style/index.js
@@ -1,0 +1,2 @@
+export { default as GlobalStyle } from './globalStyle';
+export { default as theme } from './theme';
